Fileexplorer use return value to update directory
When uploading a directory, client was asking directly to list container to update its list.
This may cause trouble in a clustered env where files hadn't time to propagate to other servers
thus responding with a list not yet up to date. This also removes one server round-trip.

This change the ordering. New items are at the end (not sorted) until the list is refreshed